+++ ui/src/main/java/com/fibelatti/ui/foundation/Density.kt
@@ -7,7 +7,10 @@ import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.unit.Dp
 
 @Composable
-fun Dp.dpToPx() = with(LocalDensity.current) { this@dpToPx.toPx() }
+fun Int.dpToPx(): Float = with(LocalDensity.current) { this@dpToPx * density }
 
 @Composable
-fun Int.pxToDp() = with(LocalDensity.current) { this@pxToDp.toDp() }
+fun Dp.dpToPx(): Float = with(LocalDensity.current) { this@dpToPx.toPx() }
+
+@Composable
+fun Int.pxToDp(): Dp = with(LocalDensity.current) { this@pxToDp.toDp() }